+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 - optimize pprint
 - fix split with ifs;
 - fix match as cmd;
+- fix pprint modules hint.
+type `String. ` to see pretty table hint.
+
 
 ## [0.6.9]
 - highlight builtin cmds;